# openapi.yaml
openapi: 3.0.0
info:
  title: Snake Arena API
  description: API for the Snake Arena game, based on the frontend mock backend.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        highScore:
          type: integer
      required:
        - id
        - username
        - email
        - highScore

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [pass-through, walls]
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - username
        - score
        - mode
        - timestamp

    GameSession:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [pass-through, walls]
        isActive:
          type: boolean
      required:
        - id
        - userId
        - username
        - score
        - mode
        - isActive

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required:
        - user
        - token

paths:
  /auth/signup:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
              required:
                - username
                - email
                - password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful

  /auth/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  /leaderboard:
    get:
      summary: Get leaderboard entries
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [pass-through, walls]
          description: Filter by game mode
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Limit number of entries
      responses:
        '200':
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'

    post:
      summary: Submit a score
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                mode:
                  type: string
                  enum: [pass-through, walls]
              required:
                - score
                - mode
      responses:
        '200':
          description: Score submitted successfully
        '401':
          description: Not authenticated

  /sessions:
    get:
      summary: Get active game sessions
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameSession'

    post:
      summary: Create a new game session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [pass-through, walls]
              required:
                - mode
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        '401':
          description: Not authenticated

  /sessions/{id}:
    get:
      summary: Get a game session by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSession'
        '404':
          description: Session not found

    put:
      summary: Update session score
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
              required:
                - score
      responses:
        '200':
          description: Session updated successfully
        '404':
          description: Session not found

    delete:
      summary: End a game session
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session ended successfully
        '404':
          description: Session not found